requirements.txt
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pexpect==4.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sos==4.9.2
ssh-import-id==5.11
systemd-python==235
Twisted==24.3.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
xkit==0.0.0
zope.interface==6.1
litellm>=1.40.0
python-telegram-bot>=20.7
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=2.0.0
bm25s>=0.2.0
ddgs>=9.0
curl_cffi>=0.7.0

telegram~=0.0.1
sentence-transformers>=2.2.0
chromadb>=0.4.0

# External packages (own repos)
yazbunu @ git+https://github.com/msakirc/yazbunu.git

# Local packages
-e ./packages/vecihi
-e ./packages/yasar_usta
-e ./packages/dallama
-e ./packages/nerd_herd
-e ./packages/kuleden_donen_var
-e ./packages/hallederiz_kadir
-e ./packages/fatih_hoca
-e ./packages/salako
-e ./packages/general_beckman

# Test dependencies
pytest-aiohttp>=1.0.0
